[Refactor][Deprecate] Deprecate karel.js
